Make AudioItem and VideoItem subclasses of new MediaItem class
- MediaItem is subclass of BaseItem
- AudioItem used if audio only playback is selected
