+++ resources/lib/youtube_plugin/kodion/constants/const_content_types.py
@@ -12,6 +12,8 @@ from __future__ import absolute_import, division, unicode_literals
 
 VIDEO_CONTENT = 'videos'
 LIST_CONTENT = 'files'
+
+AUDIO_TYPE = 'music'
 VIDEO_TYPE = 'video'
 
 FILES = 'files'
