+++ resources/lib/youtube_plugin/kodion/items/xbmc/xbmc_items.py
@@ -28,62 +28,31 @@ from ...utils import current_system_version, datetime_parser, redact_ip
 
 
 def set_info(list_item, item, properties, set_play_count=True, resume=True):
-    is_video = False
     if not current_system_version.compatible(20, 0):
         if isinstance(item, VideoItem):
-            is_video = True
             info_labels = {}
+            info_type = 'video'
 
             value = item.get_aired(as_info_label=True)
             if value is not None:
                 info_labels['aired'] = value
-
-            value = item.get_artists()
-            if value is not None:
-                info_labels['artist'] = value
 
             value = item.get_cast()
             if value is not None:
                 info_labels['castandrole'] = [(member['name'], member['role'])
                                               for member in value]
 
-            value = item.get_code()
+            value = item.get_production_code()
             if value is not None:
                 info_labels['code'] = value
-
-            value = item.get_count()
-            if value is not None:
-                info_labels['count'] = value
-
-            value = item.get_date(as_info_label=True)
-            if value is not None:
-                info_labels['date'] = value
 
             value = item.get_dateadded(as_info_label=True)
             if value is not None:
                 info_labels['dateadded'] = value
 
-            value = item.get_duration()
-            if value is not None:
-                info_labels['duration'] = value
-
             value = item.get_episode()
             if value is not None:
                 info_labels['episode'] = value
-
-            value = item.get_last_played(as_info_label=True)
-            if value is not None:
-                info_labels['lastplayed'] = value
-
-            value = item.get_mediatype()
-            if value is not None:
-                info_labels['mediatype'] = value
-
-            value = item.get_play_count()
-            if value is not None:
-                if set_play_count:
-                    info_labels['playcount'] = value
-                properties[PLAY_COUNT] = value
 
             value = item.get_plot()
             if value is not None:
@@ -93,35 +62,29 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
             if value is not None:
                 info_labels['premiered'] = value
 
-            value = item.get_rating()
-            if value is not None:
-                info_labels['rating'] = value
-
             value = item.get_season()
             if value is not None:
                 info_labels['season'] = value
-
-            value = item.get_title()
-            if value is not None:
-                info_labels['title'] = value
-
-            value = item.get_track_number()
-            if value is not None:
-                info_labels['tracknumber'] = value
-
-            value = item.get_year()
-            if value is not None:
-                info_labels['year'] = value
 
             value = item.get_studios()
             if value is not None:
                 info_labels['studio'] = value
 
-            if info_labels:
-                list_item.setInfo('video', info_labels)
+        elif isinstance(item, AudioItem):
+            info_labels = {}
+            info_type = 'music'
+
+            value = item.get_album_name()
+            if value is not None:
+                info_labels['album'] = value
+
+            value = item.get_plot()
+            if value is not None:
+                info_labels['plot'] = value
 
         elif isinstance(item, DirectoryItem):
             info_labels = {}
+            info_type = 'video'
 
             value = item.get_name()
             if value is not None:
@@ -132,45 +95,11 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
                 info_labels['plot'] = value
 
             if info_labels:
-                list_item.setInfo('video', info_labels)
+                list_item.setInfo(info_type, info_labels)
 
             if properties:
                 list_item.setProperties(properties)
             return
-
-        elif isinstance(item, AudioItem):
-            info_labels = {}
-
-            value = item.get_album_name()
-            if value is not None:
-                info_labels['album'] = value
-
-            value = item.get_artists()
-            if value is not None:
-                info_labels['artist'] = value
-
-            value = item.get_duration()
-            if value is not None:
-                info_labels['duration'] = value
-
-            value = item.get_rating()
-            if value is not None:
-                info_labels['rating'] = value
-
-            value = item.get_title()
-            if value is not None:
-                info_labels['title'] = value
-
-            value = item.get_track_number()
-            if value is not None:
-                info_labels['tracknumber'] = value
-
-            value = item.get_year()
-            if value is not None:
-                info_labels['year'] = value
-
-            if info_labels:
-                list_item.setInfo('music', info_labels)
 
         elif isinstance(item, ImageItem):
             value = item.get_title()
@@ -181,28 +110,78 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
                 list_item.setProperties(properties)
             return
 
+        else:
+            return
+
+        value = item.get_artists_string()
+        if value is not None:
+            info_labels['artist'] = value
+
+        value = item.get_count()
+        if value is not None:
+            info_labels['count'] = value
+
+        value = item.get_date(as_info_label=True)
+        if value is not None:
+            info_labels['date'] = value
+
+        value = item.get_duration()
+        if value is not None:
+            info_labels['duration'] = value
+
+        value = item.get_last_played(as_info_label=True)
+        if value is not None:
+            info_labels['lastplayed'] = value
+
+        value = item.get_mediatype()
+        if value is not None:
+            info_labels['mediatype'] = value
+
+        value = item.get_play_count()
+        if value is not None:
+            if set_play_count:
+                info_labels['playcount'] = value
+            properties[PLAY_COUNT] = value
+
+        value = item.get_rating()
+        if value is not None:
+            info_labels['rating'] = value
+
+        value = item.get_title()
+        if value is not None:
+            info_labels['title'] = value
+
+        value = item.get_track_number()
+        if value is not None:
+            info_labels['tracknumber'] = value
+
+        value = item.get_year()
+        if value is not None:
+            info_labels['year'] = value
+
         resume_time = resume and item.get_start_time()
         if resume_time:
             properties['ResumeTime'] = str(resume_time)
         duration = item.get_duration()
         if duration:
             properties['TotalTime'] = str(duration)
-            if is_video:
-                list_item.addStreamInfo('video', {'duration': duration})
+            if info_type == 'video':
+                list_item.addStreamInfo(info_type, {'duration': duration})
 
         if properties:
             list_item.setProperties(properties)
+
+        if info_labels:
+            list_item.setInfo(info_type, info_labels)
         return
 
     value = item.get_date(as_info_label=True)
     if value is not None:
         list_item.setDateTime(value)
 
-    info_tag = None
-
     if isinstance(item, VideoItem):
-        is_video = True
         info_tag = list_item.getVideoInfoTag()
+        info_type = 'video'
 
         value = item.get_aired(as_info_label=True)
         if value is not None:
@@ -212,13 +191,16 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
         if value is not None:
             info_tag.setDateAdded(value)
 
-        value = item.get_last_played(as_info_label=True)
-        if value is not None:
-            info_tag.setLastPlayed(value)
-
         value = item.get_premiered(as_info_label=True)
         if value is not None:
             info_tag.setPremiered(value)
+
+        # artist: list[str]
+        # eg. ["Angerfist"]
+        # Used as alias for channel name
+        value = item.get_artists()
+        if value is not None:
+            info_tag.setArtists(value)
 
         # cast: list[xbmc.Actor]
         # From list[{member: str, role: str, order: int, thumbnail: str}]
@@ -227,26 +209,12 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
         if value is not None:
             info_tag.setCast([xbmc.Actor(**member) for member in value])
 
-        # code: str
-        # eg. "466K | 3.9K | 312"
-        # Production code, currently used to store misc video data for label
-        # formatting
-        value = item.get_code()
-        if value is not None:
-            info_tag.setProductionCode(value)
-
-        # count: int
-        # eg. 12
-        # Can be used to store an id for later, or for sorting purposes
-        # Used for Youtube video view count
-        value = item.get_count()
-        if value is not None:
-            list_item.setInfo('video', {'count': value})
-
         # director: list[str]
         # eg. "Steven Spielberg"
         # Currently unused
-        # info_tag.setDirectors(item.get_directors())
+        # value = item.get_directors()
+        # if value is not None:
+        #     info_tag.setDirectors(value)
 
         # episode: int
         value = item.get_episode()
@@ -256,24 +224,22 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
         # imdbnumber: str
         # eg. "tt3458353"
         # Currently unused
-        # info_tag.setIMDBNumber(item.get_imdb_id())
-
-        # mediatype: str
-        value = item.get_mediatype()
-        if value is not None:
-            info_tag.setMediaType(value)
-
-        # playcount: int
-        value = item.get_play_count()
-        if value is not None:
-            if set_play_count:
-                info_tag.setPlaycount(value)
-            properties[PLAY_COUNT] = value
+        # value = item.get_imdb_id()
+        # if value is not None:
+        #     info_tag.setIMDBNumber(value)
 
         # plot: str
         value = item.get_plot()
         if value is not None:
             info_tag.setPlot(value)
+
+        # code: str
+        # eg. "466K | 3.9K | 312"
+        # Production code, currently used to store misc video data for label
+        # formatting
+        value = item.get_production_code()
+        if value is not None:
+            info_tag.setProductionCode(value)
 
         # season: int
         value = item.get_season()
@@ -285,6 +251,44 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
         value = item.get_studios()
         if value is not None:
             info_tag.setStudios(value)
+
+        # tracknumber: int
+        # eg. 12
+        value = item.get_track_number()
+        if value is not None:
+            info_tag.setTrackNumber(value)
+
+    elif isinstance(item, AudioItem):
+        info_tag = list_item.getMusicInfoTag()
+        info_type = 'music'
+
+        value = item.get_premiered(as_info_label=True)
+        if value is not None:
+            info_tag.setReleaseDate(value)
+
+        # album: str
+        # eg. "Buckle Up"
+        value = item.get_album_name()
+        if value is not None:
+            info_tag.setAlbum(value)
+
+        # artist: str
+        # eg. "Artist 1, Artist 2"
+        # Used as alias for channel name
+        value = item.get_artists_string()
+        if value is not None:
+            info_tag.setArtist(value)
+
+        # comment: str
+        value = item.get_plot()
+        if value is not None:
+            info_tag.setComment(value)
+
+        # track: int
+        # eg. 12
+        value = item.get_track_number()
+        if value is not None:
+            info_tag.setTrack(value)
 
     elif isinstance(item, DirectoryItem):
         info_tag = list_item.getVideoInfoTag()
@@ -312,14 +316,8 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
             list_item.setProperties(properties)
         return
 
-    elif isinstance(item, AudioItem):
-        info_tag = list_item.getMusicInfoTag()
-
-        # album: str
-        # eg. "Buckle Up"
-        value = item.get_album_name()
-        if value is not None:
-            info_tag.setAlbum(value)
+    else:
+        return
 
     resume_time = resume and item.get_start_time()
     duration = item.get_duration()
@@ -327,25 +325,44 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
         info_tag.setResumePoint(resume_time, float(duration))
     elif resume_time:
         info_tag.setResumePoint(resume_time)
-    if is_video and duration:
+    if info_type == 'video' and duration:
         info_tag.addVideoStream(xbmc.VideoStreamDetail(duration=duration))
-
-    # artist: list[str]
-    # eg. ["Angerfist"]
-    # Used as alias for channel name
-    value = item.get_artists()
-    if value is not None:
-        info_tag.setArtists(value)
 
     # duration: int
     # As seconds
     if duration is not None:
         info_tag.setDuration(duration)
 
+    # mediatype: str
+    value = item.get_mediatype()
+    if value is not None:
+        info_tag.setMediaType(value)
+
+    value = item.get_last_played(as_info_label=True)
+    if value is not None:
+        info_tag.setLastPlayed(value)
+
+    # playcount: int
+    value = item.get_play_count()
+    if value is not None:
+        if set_play_count:
+            info_tag.setPlaycount(value)
+        properties[PLAY_COUNT] = value
+
+    # count: int
+    # eg. 12
+    # Can be used to store an id for later, or for sorting purposes
+    # Used for Youtube video view count
+    value = item.get_count()
+    if value is not None:
+        list_item.setInfo(info_type, {'count': value})
+
     # genre: list[str]
     # eg. ["Hardcore"]
     # Currently unused
-    # info_tag.setGenres(item.get_genres())
+    # value = item.get_genres()
+    # if value is not None:
+    #     info_tag.setGenres(value)
 
     # rating: float
     value = item.get_rating()
@@ -358,12 +375,6 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
     if value is not None:
         info_tag.setTitle(value)
 
-    # tracknumber: int
-    # eg. 12
-    value = item.get_track_number()
-    if value is not None:
-        info_tag.setTrackNumber(value)
-
     # year: int
     # eg. 1994
     value = item.get_year()
@@ -374,14 +385,12 @@ def set_info(list_item, item, properties, set_play_count=True, resume=True):
         list_item.setProperties(properties)
 
 
-def video_playback_item(context, video_item, show_fanart=None, **_kwargs):
-    uri = video_item.get_uri()
-    context.log_debug('Converting VideoItem |%s|' % redact_ip(uri))
+def playback_item(context, media_item, show_fanart=None, **_kwargs):
+    uri = media_item.get_uri()
+    context.log_debug('Converting %s |%s|' % (media_item.__class__.__name__,
+                                              redact_ip(uri)))
 
     settings = context.get_settings()
-    headers = video_item.get_headers()
-    license_key = video_item.get_license_key()
-
     ui = context.get_ui()
     is_external = ui.get_property(PLAY_WITH)
     is_strm = context.get_param('strm')
@@ -395,19 +404,19 @@ def video_playback_item(context, video_item, show_fanart=None, **_kwargs):
         props = {}
     else:
         kwargs = {
-            'label': video_item.get_title() or video_item.get_name(),
-            'label2': video_item.get_short_details(),
+            'label': media_item.get_title() or media_item.get_name(),
+            'label2': media_item.get_short_details(),
             'path': uri,
             'offscreen': True,
         }
         props = {
-            'isPlayable': str(video_item.playable).lower(),
+            'isPlayable': str(media_item.playable).lower(),
         }
 
-    if video_item.use_isa_video() and context.use_inputstream_adaptive():
+    if media_item.use_isa() and context.use_inputstream_adaptive():
         capabilities = context.inputstream_adaptive_capabilities()
 
-        use_mpd = video_item.use_mpd_video()
+        use_mpd = media_item.use_mpd()
         if use_mpd:
             manifest_type = 'mpd'
             mime_type = 'application/dash+xml'
@@ -430,7 +439,7 @@ def video_playback_item(context, video_item, show_fanart=None, **_kwargs):
         if not current_system_version.compatible(21, 0):
             props['inputstream.adaptive.manifest_type'] = manifest_type
 
-        if video_item.live:
+        if media_item.live:
             if 'manifest_config_prop' in capabilities:
                 props['inputstream.adaptive.manifest_config'] = dumps({
                     'timeshift_bufferlimit': 4 * 60 * 60,
@@ -443,10 +452,12 @@ def video_playback_item(context, video_item, show_fanart=None, **_kwargs):
                 'ssl_verify_peer': False,
             })
 
+        headers = media_item.get_headers()
         if headers:
             props['inputstream.adaptive.manifest_headers'] = headers
             props['inputstream.adaptive.stream_headers'] = headers
 
+        license_key = media_item.get_license_key()
         if license_key:
             props['inputstream.adaptive.license_type'] = 'com.widevine.alpha'
             props['inputstream.adaptive.license_key'] = license_key
@@ -456,6 +467,7 @@ def video_playback_item(context, video_item, show_fanart=None, **_kwargs):
             mime_type = uri.split('mime=', 1)[1].split('&', 1)[0]
             mime_type = mime_type.replace('%2F', '/')
 
+        headers = media_item.get_headers()
         if (headers and uri.startswith('http')
                 and not (is_external
                          or settings.default_player_web_urls())):
@@ -473,64 +485,21 @@ def video_playback_item(context, video_item, show_fanart=None, **_kwargs):
 
     if show_fanart is None:
         show_fanart = settings.fanart_selection()
-    image = video_item.get_image()
+    image = media_item.get_image()
     art = {'icon': image}
     if image:
         art['thumb'] = image
     if show_fanart:
-        art['fanart'] = video_item.get_fanart()
+        art['fanart'] = media_item.get_fanart()
     list_item.setArt(art)
 
-    if video_item.subtitles:
-        list_item.setSubtitles(video_item.subtitles)
+    if media_item.subtitles:
+        list_item.setSubtitles(media_item.subtitles)
 
     resume = context.get_param('resume')
-    set_info(list_item, video_item, props, resume=resume)
+    set_info(list_item, media_item, props, resume=resume)
 
     return list_item
-
-
-def audio_listitem(context,
-                   audio_item,
-                   show_fanart=None,
-                   for_playback=False,
-                   **_kwargs):
-    uri = audio_item.get_uri()
-    context.log_debug('Converting AudioItem |%s|' % uri)
-
-    kwargs = {
-        'label': audio_item.get_title() or audio_item.get_name(),
-        'label2': audio_item.get_short_details(),
-        'path': uri,
-        'offscreen': True,
-    }
-    props = {
-        'isPlayable': str(audio_item.playable).lower(),
-        'ForceResolvePlugin': 'true',
-    }
-
-    list_item = xbmcgui.ListItem(**kwargs)
-
-    if show_fanart is None:
-        show_fanart = context.get_settings().fanart_selection()
-    image = audio_item.get_image()
-    art = {'icon': image}
-    if image:
-        art['thumb'] = image
-    if show_fanart:
-        art['fanart'] = audio_item.get_fanart()
-    list_item.setArt(art)
-
-    resume = context.get_param('resume') or not for_playback
-    set_info(list_item, audio_item, props, resume=resume)
-
-    context_menu = audio_item.get_context_menu()
-    if context_menu:
-        list_item.addContextMenuItems(context_menu)
-
-    if for_playback:
-        return list_item
-    return uri, list_item, False
 
 
 def directory_listitem(context, directory_item, show_fanart=None, **_kwargs):
@@ -655,33 +624,34 @@ def uri_listitem(context, uri_item, **_kwargs):
     return list_item
 
 
-def video_listitem(context,
-                   video_item,
+def media_listitem(context,
+                   media_item,
                    show_fanart=None,
                    focused=None,
                    **_kwargs):
-    uri = video_item.get_uri()
-    context.log_debug('Converting VideoItem |%s|' % uri)
+    uri = media_item.get_uri()
+    context.log_debug('Converting %s |%s|' % (media_item.__class__.__name__,
+                                              uri))
 
     kwargs = {
-        'label': video_item.get_title() or video_item.get_name(),
-        'label2': video_item.get_short_details(),
+        'label': media_item.get_title() or media_item.get_name(),
+        'label2': media_item.get_short_details(),
         'path': uri,
         'offscreen': True,
     }
     props = {
-        'isPlayable': str(video_item.playable).lower(),
+        'isPlayable': str(media_item.playable).lower(),
         'ForceResolvePlugin': 'true',
     }
 
-    published_at = video_item.get_added_utc()
-    scheduled_start = video_item.get_scheduled_start_utc()
+    published_at = media_item.get_added_utc()
+    scheduled_start = media_item.get_scheduled_start_utc()
     datetime = scheduled_start or published_at
     local_datetime = None
     if datetime:
         local_datetime = datetime_parser.utc_to_local(datetime)
         props['PublishedLocal'] = to_str(local_datetime)
-    if video_item.live:
+    if media_item.live:
         props['PublishedSince'] = context.localize('live')
     elif local_datetime:
         props['PublishedSince'] = to_str(datetime_parser.datetime_to_since(
@@ -690,7 +660,7 @@ def video_listitem(context,
 
     set_play_count = True
     resume = True
-    prop_value = video_item.video_id
+    prop_value = media_item.video_id
     if prop_value:
         if focused and focused == prop_value:
             set_play_count = False
@@ -698,22 +668,22 @@ def video_listitem(context,
         props[VIDEO_ID] = prop_value
 
     # make channel_id property available for keymapping
-    prop_value = video_item.channel_id
+    prop_value = media_item.channel_id
     if prop_value:
         props[CHANNEL_ID] = prop_value
 
     # make subscription_id property available for keymapping
-    prop_value = video_item.subscription_id
+    prop_value = media_item.subscription_id
     if prop_value:
         props[SUBSCRIPTION_ID] = prop_value
 
     # make playlist_id property available for keymapping
-    prop_value = video_item.playlist_id
+    prop_value = media_item.playlist_id
     if prop_value:
         props[PLAYLIST_ID] = prop_value
 
     # make playlist_item_id property available for keymapping
-    prop_value = video_item.playlist_item_id
+    prop_value = media_item.playlist_item_id
     if prop_value:
         props[PLAYLISTITEM_ID] = prop_value
 
@@ -721,34 +691,34 @@ def video_listitem(context,
 
     if show_fanart is None:
         show_fanart = context.get_settings().fanart_selection()
-    image = video_item.get_image()
+    image = media_item.get_image()
     art = {'icon': image}
     if image:
         art['thumb'] = image
     if show_fanart:
-        art['fanart'] = video_item.get_fanart()
+        art['fanart'] = media_item.get_fanart()
     list_item.setArt(art)
 
-    if video_item.subtitles:
-        list_item.setSubtitles(video_item.subtitles)
+    if media_item.subtitles:
+        list_item.setSubtitles(media_item.subtitles)
 
     set_info(list_item,
-             video_item,
+             media_item,
              props,
              set_play_count=set_play_count,
              resume=resume)
 
     if not set_play_count:
-        video_id = video_item.video_id
+        video_id = media_item.video_id
         playback_history = context.get_playback_history()
         playback_history.set_item(video_id, dict(
             playback_history.get_item(video_id) or {},
-            play_count=int(not video_item.get_play_count()),
+            play_count=int(not media_item.get_play_count()),
             played_time=0.0,
             played_percent=0,
         ))
 
-    context_menu = video_item.get_context_menu()
+    context_menu = media_item.get_context_menu()
     if context_menu:
         list_item.addContextMenuItems(context_menu)
