+++ resources/lib/youtube_plugin/youtube/helper/yt_play.py
@@ -29,7 +29,7 @@ from ...kodion.constants import (
     SERVER_POST_START,
     SERVER_WAKEUP,
 )
-from ...kodion.items import VideoItem
+from ...kodion.items import AudioItem, VideoItem
 from ...kodion.network import get_connect_address
 from ...kodion.utils import find_video_id, select_stream
 
@@ -119,13 +119,16 @@ def _play_stream(provider, context):
         ))
         stream['url'] = url
 
-    video_item = VideoItem(metadata.get('title', ''), stream['url'])
+    if audio_only or not video_type:
+        media_item = AudioItem(metadata.get('title', ''), stream['url'])
+    else:
+        media_item = VideoItem(metadata.get('title', ''), stream['url'])
 
     use_history = not (screensaver or incognito or stream.get('live'))
     use_remote_history = use_history and settings.use_remote_history()
     use_play_data = use_history and settings.use_local_history()
 
-    utils.update_play_info(provider, context, video_id, video_item,
+    utils.update_play_info(provider, context, video_id, media_item,
                            stream, use_play_data=use_play_data)
 
     seek_time = 0.0 if params.get('resume') else params.get('seek', 0.0)
@@ -133,20 +136,20 @@ def _play_stream(provider, context):
     end_time = params.get('end', 0.0)
 
     if start_time:
-        video_item.set_start_time(start_time)
+        media_item.set_start_time(start_time)
     # Setting the duration based on end_time can cause issues with
     # listing/sorting and other addons that monitor playback
     # if end_time:
     #     video_item.set_duration_from_seconds(end_time)
 
-    play_count = use_play_data and video_item.get_play_count() or 0
+    play_count = use_play_data and media_item.get_play_count() or 0
     playback_stats = stream.get('playback_stats')
 
     playback_data = {
         'video_id': video_id,
         'channel_id': metadata.get('channel', {}).get('id', ''),
         'video_status': metadata.get('status', {}),
-        'playing_file': video_item.get_uri(),
+        'playing_file': media_item.get_uri(),
         'play_count': play_count,
         'use_remote_history': use_remote_history,
         'use_local_history': use_play_data,
@@ -160,7 +163,7 @@ def _play_stream(provider, context):
 
     ui.set_property(PLAYER_DATA, json.dumps(playback_data, ensure_ascii=False))
     context.send_notification(PLAYBACK_INIT, playback_data)
-    return video_item
+    return media_item
 
 
 def _play_playlist(provider, context):
@@ -342,9 +345,9 @@ def process(provider, context, **_kwargs):
             return False
         ui.clear_property(SERVER_POST_START)
         context.wakeup(SERVER_WAKEUP, timeout=5)
-        video_item = _play_stream(provider, context)
+        media_item = _play_stream(provider, context)
         ui.set_property(SERVER_POST_START)
-        return video_item
+        return media_item
 
     if playlist_id or 'playlist_ids' in params:
         return _play_playlist(provider, context)
